Fix to correctly handle stop words in queries.
Stop words were being created and compared without case folding. Thus
the expected behaviour to remove all instance of "The, the,..." or other
such example was not done.

Now stopwords are casefolded and digits are removed when creating the list.
Now stopwords are casefolded when compressing the index.

Handling of word removal filters (No numbers and stopwords) during evaluation
was not done properly. Any stop words with a conjunction resulted in no
results. The expected behaviour was to disregard those terms.